move ignore errors serialization tests to kernel tests
