+++ tensorflow/python/data/experimental/kernel_tests/serialization/BUILD
@@ -222,26 +222,6 @@ tf_py_test(
 )
 
 tf_py_test(
-    name = "ignore_errors_serialization_test",
-    size = "small",
-    srcs = ["ignore_errors_serialization_test.py"],
-    tags = [
-        "no_oss",
-        "no_pip",
-        "no_windows",
-        "noasan",  # TODO(b/337374867) fails with -fsanitize=null
-    ],
-    deps = [
-        "//tensorflow/python:array_ops",
-        "//tensorflow/python:client_testlib",
-        "//tensorflow/python/data/experimental/ops:error_ops",
-        "//tensorflow/python/data/kernel_tests:checkpoint_test_base",
-        "//tensorflow/python/data/ops:dataset_ops",
-        "//third_party/py/numpy",
-    ],
-)
-
-tf_py_test(
     name = "interleave_dataset_serialization_test",
     size = "medium",
     srcs = ["interleave_dataset_serialization_test.py"],
